Implemented US10
US10:As an enemy saucer I want to my body slam be minus 1 health to player health [15 mins]

+++ Week4_SpaceGame/Assets/Em-Scripts/PlayerHealth.cs
@@ -18,7 +18,7 @@ public class PlayerHealth : MonoBehaviour
         
     }
 
-    public void BeenHit(int livesLost)
+    public void BeenHit(float livesLost)
     {
         currentHealth -= livesLost;
         if(currentHealth <= 0)
